simplify args in main
