+++ xinstall/ide.py
@@ -292,3 +292,14 @@ def _add_subparser_intellij_idea_scala(subparsers) -> None:
     add_subparser(
         subparsers, "IntelliJ IDEA", func=intellij_idea, aliases=["intellij", "idea"]
     )
+
+
+def _add_subparser_ide(subparsers):
+    _add_subparser_vim(subparsers)
+    _add_subparser_neovim(subparsers)
+    _add_subparser_spacevim(subparsers)
+    _add_subparser_ideavim(subparsers)
+    _add_subparser_visual_studio_code(subparsers)
+    _add_subparser_intellij_idea_scala(subparsers)
+    _add_subparser_bash_lsp(subparsers)
+
